feat: add useQueryReminders
