+++ examples/SampleApp/src/components/Reminders/RemindersList.tsx
@@ -1,14 +1,8 @@
 import { useCallback, useEffect, useState } from 'react';
 import { FlatList, Pressable, ScrollView, StyleSheet, Text, View } from 'react-native';
-import { useChatContext, useStateStore, useTheme } from 'stream-chat-react-native';
-import { Event, PaginatorState, ReminderResponse } from 'stream-chat';
+import { useChatContext, useTheme, useQueryReminders } from 'stream-chat-react-native';
+import { ReminderResponse } from 'stream-chat';
 import { ReminderItem } from './ReminderItem';
-
-const selector = (nextValue: PaginatorState<ReminderResponse>) =>
-  ({
-    isLoading: nextValue.isLoading,
-    reminders: nextValue.items,
-  }) as const;
 
 const tabs = [
   { key: 'all', title: 'All' },
@@ -23,24 +17,7 @@ type TabItemType = {
   title: string;
 };
 
-// Utility to sort reminders by remind_at date in ascending order
-const sortRemindersByDate = (reminders: ReminderResponse[]) => {
-  return reminders.sort((a, b) => {
-    if (!a.remind_at || !b.remind_at) {
-      return 0; // If either remind_at is missing, keep original order
-    }
-    // Sort by remind_at date
-    return new Date(a.remind_at).getTime() - new Date(b.remind_at).getTime();
-  });
-};
-
-const isReminderOverdue = (reminder?: ReminderResponse) => {
-  return reminder?.remind_at && new Date(reminder.remind_at) < new Date();
-};
-
-const isReminderUpcoming = (reminder?: ReminderResponse) => {
-  return reminder?.remind_at && new Date(reminder.remind_at) > new Date();
-};
+const renderItem = ({ item }: { item: ReminderResponse }) => <ReminderItem {...item} />;
 
 export const RemindersList = () => {
   const [selectedTab, setSelectedTab] = useState<TabItemType>(tabs[0]);
@@ -50,84 +27,13 @@ export const RemindersList = () => {
     },
   } = useTheme();
   const { client } = useChatContext();
-  const { isLoading, reminders } = useStateStore(client.reminders.paginator.state, selector);
-  const [data, setData] = useState<ReminderResponse[]>(reminders ?? []);
+
+  const { data, isLoading, onEndReached, onRefresh } = useQueryReminders();
 
   useEffect(() => {
+    client.reminders.paginator.filters = {};
     client.reminders.paginator.sort = { remind_at: 1 };
-    client.reminders.paginator.state.subscribeWithSelector(
-      ({ items }) => [items],
-      ([items]) => {
-        setData(items ?? []);
-      },
-    );
   }, [client.reminders.paginator]);
-
-  useEffect(() => {
-    const handleReminderDeleted = (event: Event) => {
-      if (!event.reminder?.message_id) {
-        return;
-      }
-      setData((prevData) => {
-        return prevData.filter((item) => item.message_id !== event.reminder?.message_id);
-      });
-    };
-
-    const handleReminderCreated = (event: Event) => {
-      setData((prevData) => {
-        if (!event.reminder) {
-          return prevData;
-        }
-        const updatedData = [...prevData, event.reminder];
-        return sortRemindersByDate(updatedData);
-      });
-    };
-
-    const handleReminderUpdated = (event: Event) => {
-      if (selectedTab.key === 'all') {
-        return;
-      }
-      const { reminder } = event;
-      setData((prevData) => {
-        if (!reminder) {
-          return prevData; // No update needed if reminder is undefined
-        }
-        const existingReminder = prevData.find((item) => item.message_id === reminder?.message_id);
-        if (!existingReminder) {
-          return prevData; // No update needed if reminder not found
-        }
-
-        if (existingReminder.remind_at && !event.reminder?.remind_at) {
-          return prevData.filter((item) => item.message_id !== event.reminder?.message_id);
-        }
-        if (!existingReminder.remind_at && event.reminder?.remind_at) {
-          return prevData.filter((item) => item.message_id !== event.reminder?.message_id);
-        }
-        if (isReminderOverdue(existingReminder) && !isReminderOverdue(event.reminder)) {
-          return prevData.filter((item) => item.message_id !== event.reminder?.message_id);
-        }
-        if (isReminderUpcoming(existingReminder) && !isReminderUpcoming(event.reminder)) {
-          return prevData.filter((item) => item.message_id !== event.reminder?.message_id);
-        }
-
-        return prevData;
-      });
-    };
-
-    const listeners = [
-      client.on('reminder.created', handleReminderCreated),
-      client.on('reminder.deleted', handleReminderDeleted),
-      client.on('reminder.updated', handleReminderUpdated),
-    ];
-
-    return () => {
-      listeners.forEach((l) => l.unsubscribe());
-    };
-  }, [client, selectedTab]);
-
-  const onEndReached = useCallback(() => {
-    client.reminders.queryNextReminders();
-  }, [client.reminders]);
 
   const onChangeTab = useCallback(
     async (tab: TabItemType) => {
@@ -154,15 +60,6 @@ export const RemindersList = () => {
       await client.reminders.queryNextReminders();
     },
     [client.reminders],
-  );
-
-  const onRefresh = useCallback(async () => {
-    await client.reminders.queryNextReminders();
-  }, [client.reminders]);
-
-  const renderItem = useCallback(
-    ({ item }: { item: ReminderResponse }) => <ReminderItem {...item} />,
-    [],
   );
 
   const renderEmptyComponent = useCallback(
