+++ package/src/hooks/index.ts
@@ -6,3 +6,4 @@ export * from './useStateStore';
 export * from './useStableCallback';
 export * from './useLoadingImage';
 export * from './useMessageReminder';
+export * from './useQueryReminders';
